Initial pass at adding API support for Processings
