+++ lib/cropster.rb
@@ -7,6 +7,7 @@ require "cropster/version"
 require "cropster/client"
 require 'cropster/constants'
 require 'cropster/lot'
+require 'cropster/processing'
 
 require 'cropster/response/response_handler'
 require 'cropster/response/formatted_response_item'
@@ -16,6 +17,6 @@ require 'cropster/response/price'
 require 'cropster/response/project'
 require 'cropster/response/source'
 require 'cropster/response/lot'
-require 'cropster/response/roast_batch'
+require 'cropster/response/processing'
 
 ServiceUnavailableError = Class.new(StandardError)
